+++ builds/compromise.js
@@ -1,4 +1,4 @@
-/* compromise 13.9.0 MIT */
+/* compromise 13.9.1 MIT */
 (function (global, factory) {
   typeof exports === 'object' && typeof module !== 'undefined' ? module.exports = factory() :
   typeof define === 'function' && define.amd ? define(factory) :
@@ -1558,6 +1558,13 @@
           implicit: true,
           reduced: true
         };
+      } else if (options === 'implicit') {
+        options = {
+          punctuation: true,
+          implicit: true,
+          whitespace: true,
+          trim: true
+        };
       } else if (options === 'root') {
         options = {
           titlecase: false,
@@ -1582,6 +1589,11 @@
     }
 
     var text = terms.reduce(function (str, t, i) {
+      // don't output intro space for a contraction-match  i'm good => "[am] good"
+      if (i === 0 && t.text === '' && t.implicit !== null && !options.implicit) {
+        return str;
+      }
+
       options.last = isLast && i === terms.length - 1;
       var showPre = true;
       var showPost = true;
@@ -1598,7 +1610,8 @@
         }
       }
 
-      var txt = t.textOut(options, showPre, showPost); // if (options.titlecase && i === 0) {
+      var txt = t.textOut(options, showPre, showPost); // console.log(terms)
+      // if (options.titlecase && i === 0) {
       // txt = titleCase(txt)
       // }
 
@@ -1976,6 +1989,10 @@
 
     if (options.reduced) {
       res.reduced = this.text('reduced');
+    }
+
+    if (options.implicit) {
+      res.implicit = this.text('implicit');
     }
 
     if (options.root) {
@@ -2480,7 +2497,13 @@
 
 
       if (state.terms[state.t].isImplicit() && regs[state.r - 1] && state.terms[state.t + 1]) {
+        // if the last match was implicit too, we're missing a word.
+        if (state.terms[state.t - 1] && state.terms[state.t - 1].implicit === regs[state.r - 1].word) {
+          return null;
+        } // console.log(state.terms[state.t])
         // does the next one match?
+
+
         if (state.terms[state.t + 1].doesMatch(reg, state.start_i + state.t, state.phrase_length)) {
           state.t += 2;
           continue;
@@ -3796,7 +3819,7 @@
 
   var fromJSON_1 = fromJSON;
 
-  var _version = '13.9.0';
+  var _version = '13.9.1';
 
   var _data = {
     "Comparative": "true¦better",
@@ -8398,7 +8421,7 @@
     var list = [].concat(this.list);
     var obj = {};
     list = list.filter(function (p) {
-      var str = p.text('reduced').trim();
+      var str = p.text('reduced').trim() || p.text('implicit').trim();
 
       if (obj.hasOwnProperty(str) === true) {
         return false;
@@ -13011,9 +13034,9 @@
 
       _createClass(Lists, [{
         key: "conjunctions",
-
+        value:
         /** coordinating conjunction */
-        value: function conjunctions() {
+        function conjunctions() {
           return this.match('(and|or)');
         }
         /** split-up by list object */
@@ -13493,9 +13516,9 @@
 
       _createClass(Parentheses, [{
         key: "unwrap",
-
+        value:
         /** remove the parentheses characters */
-        value: function unwrap() {
+        function unwrap() {
           this.list.forEach(function (p) {
             var first = p.terms(0);
             first.pre = first.pre.replace(open, '');
@@ -13667,9 +13690,9 @@
 
       _createClass(Quotations, [{
         key: "unwrap",
-
+        value:
         /** remove the quote characters */
-        value: function unwrap() {
+        function unwrap() {
           return this;
         }
       }]);