+++ scratch.js
@@ -11,7 +11,35 @@ nlp.extend(require('./plugins/penn-tags/src'))
 //
 //
 
-console.log(nlp('in the town where I was born').pennTags({ offset: true })[0])
+/*
+  have  - false
+  [not] - true
+  done  - false
+*/
+
+// let doc = nlp(`haven't done`)
+// doc.match(`have done`).debug()
+
+// let doc = nlp(`i haven't done it`)
+// let m = doc.match(`not done`)
+// console.log(`|${m.text()}|`)
+// m.debug()
+
+let str = `Leonard's was Randall`
+let doc = nlp(str)
+console.log(doc.match('#Noun').text())
+
+// let doc = nlp(`is not foobar isn't`)
+// doc = doc.terms().unique().debug()
+// console.log('|' + nlp(`isn't`).text('implicit') + '|')
+
+// console.log(`|${doc.text()}|`)
+
+// doc.termList().forEach(t => {
+//   console.log(t.text, t.isImplicit())
+// })
+
+// console.log(nlp('in the town where I was born').pennTags({ offset: true })[0])
 // let reg = nlp.parseMatch(`before (#Value .) after`)
 // // console.log(JSON.stringify(reg, null, 2))
 // doc.match(reg).debug()